Update OAM w/o reboot on simplex active controller
If the active controller is a Simplex or an orphan controller, the OAM
update does not need to wait for a reboot.

This commit applies the same manifest as against the AIO-SX system on
controller-0 when updating the OAM network. Ensures the orphan
controller-0 can get the same update as AIOSX when updating the OAM
network without a reboot required.

Test plan:
1. Verify OAM update on AIOSX.
2. Verify OAM update on orphan controller-0 in a DX system, verify the
   OpenStack endpoints are assigned to the OAM IP updated, verify no
   250.001 alarm persisted.
3. Verify subcloud enrollment against a DX controller with OAM change
   succeeded after this change.

Partial-bug: 2121954

Signed-off-by: Jiang, Yuxing <Yuxing.Jiang@windriver.com>
Change-Id: I38c287ea0faf02517aaeecf4d2ab445c1fbf604d