Fix app reapply skipping the install process
When a reapply is triggered after an unsuccessful attempt to install an
app, the system was skipping the install process and directly returning
an "apply-failed" status.
This behavior was caused by a race condition between the HelmRelease
reconciliation process and the AppFrmk logic that checks the release
status.

This commit introduces a loop that waits for the HelmRelease status to
change before continuing the AppFrmk flow. The loop checks whether the
HelmRelease is in a 'Failed' status. If it is, the loop waits, since
this usually means that reconciliation hasn’t started yet. Once helm
begins reconciliation, the status changes to 'Unknown', which
indicates that the install process is progressing.

Test Plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX master fresh install.
PASS: create a bad override to invalidate some app health
      check. Apply the override with the helm-override-update command
      and install the app.
PASS: After the installation fails, remove the bad override, trigger
      the app install again, and check if the system successfully
      installs the app this time.

PASS: install an app.
PASS: create a new app tarball with a different version in the app
      metadata.
PASS: start the app update process using the new tarball and check if
      the system successfully install the app.

Closes-bug: 2121227

Change-Id: I2a1a39da438b30a141dbbb33bf9626be86ff66c2
Signed-off-by: edias <edson.dias@windriver.com>