Update kube-apiserver certificate sans on admin network update
The admin network needs to be treated just like the mgmt
network when it comes to subclouds deployed with it.

As such we need to have the admin ip address properly added
to the kube-apiserver certificate so that communication can
be done over the admin network.

Test Plan:
PASS: build-pkgs & build-image.
PASS: On an AIO-SX subcloud:
      - On a running subcloud, add the admin network
      and do dcmanager update on the system controller
      to change the subcloud management network to be
      the newly defined admin network
      - Verify that admin ip address are added to the
      kube-apiserver certificate.
      - Send curl requests from the system controller
      to subcloud as https://[SUBCLOUD-ADMIN-IP]:6443
      and verifies it reaches kube-apiserver directly.

Story: 2011399
Task: 52779

Signed-off-by: Rei Oliveira <Reinildes.JoseMateusOliveira@windriver.com>
Change-Id: I7e7fd80344eb7e0b23d9f836b252c0c91b2184cf