Fix reapply ordering
This fix moves the "dependent-apps" key to the end of the
reapplication list. This is expected behavior to ensure that
apps that are dependencies but have no dependencies are
installed first and don't break the reapplication logic.

For example:

metrics-server depends on node-feature-discovery
node-feature-discovery has no dependencies.

Both require reapplication after unlocking a node.
Node-feature-discovery will be reapplied first, and after
metrics-server will  be reapplied.

Test Plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX master fresh install.
PASS: Modified metrics-server, which depends on NFD, and both need
      to be reapplied when the platform is unlocked.
      Node-feature-discovery is reapplied before metrics-server.
PASS: Vault depends on rook-ceph, and both need to be reapplied when
      the platform is unlocked. Rook-ceph is reapplied before Vault.
PASS: stx-openstack depends on rook-ceph, and both need to be reapplied
      when the platform is unlocked. Rook-ceph is reapplied before
      stx-openstack.

Closes-bug: 2119984

Change-Id: Ia5ddebd2f65e44cd270d29b3c140e4c77d45f642
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>