Let strongswan to reauth when IPsec cert renewed
Currently when the cron job renew IPsec certificate, it will issue
command "swanctl --rekey --ike system-nodes --reauth" to reauth IPsec
SAs in order for the new cert to be used. This command is heavy, it
will try to tear down the existing SAs and establish new ones in
parallel. This could cause duplicate SPD entries and kernel refuse to
install the new policies, leading to longer time for the SAs to
establish and remain stable. In some cases this will trigger nodes
reboots.

In swanctl.conf, "reauth_time = 14400" (4 hours) is already set for
IPsec connections. So this fix removed the reauth command in
ipsec-client. When the cron job renews the IPsec cert (which happens as
early as 15 days ahead of the cert expiry date), it just reloads the
certs and connections. Instead of forcing an reauth, let the strongswan
to reauth naturally when the reauth time reached.

Test Plan:
PASS: In a DX system with IPsec SAs established, check the certficates
      in use on both controllers:
      swanctl --list-cert

      Run ipsec-client to renew IPsec certificate on both controllers:
      ipsec-client -o 2 pxecontroller

      Verify there are no reboots on both controllers.

      Check the certificate list again, verify new certificates are
      loaded (at this moment, it will show only the local host's
      cerfificate):
      swanctl --list-cert

      Wait a few hours until the strongswan auto reauth happens, check
      the certificate list again, verify the new certs are in use (it
      will show the peer's cert along with the local host's cert):
      swanctl --list-cert

PASS: In a DX system, simulate a situation where controller-1's IPsec CA
      certs are different from system-root-ca and system-local-ca certs,
      lock and unlock controller-1, verify that during rebooting
      controller_config will renew IPsec certificates, and SAs
      established on time to apply puppet manifest successfully. Once
      controller-1 fully boots up, verify IPsec SAs established, and is
      enabled and available.

Closes-Bug: 2121871
Change-Id: Ie41ee02b9f092cf935de62bc21343ee1d2d1ceb2
Signed-off-by: Andy Ning <andy.ning@windriver.com>