Do not execute admin network runtime config during migration
With the changes from 2011191 to reduce address usage in SX, the admin
runtime configuration fails to properly execute SM if in a migration
to DX scenario. This happens because SM starts to expect the unit
address (controller-0) to be configured, but until the unlock only
the floating address is configured in Linux.

This change avoid runtime execution since the system is locked and
migrating to DX. This is already done during management
reconfiguration in SX.

A second change adds the admin network to the list of networks that
need to update the address in the database. Without it the new unit
address is not associated with the desired interface-ID

Test Plan
=========

[PASS] migrate SX stand-alone to DX with admin network configured
[PASS] migrate SX subcloud to DX with admin network configured

Story: 2011191
Task: 52768

Change-Id: If388cd6749c5a64cf7b9ab7f2a80e8254bdfee71
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>