Add new force remove for host-fs ceph
In order to enable the removal of a Rook Ceph's host
that is offline, unavailable or unrecoverable a new
force option was created for the host-fs ceph only.

This functions sets a special capability "marked_for_removal"
that enables a special flow inside Rook Ceph application that
performs the remove of the components leaving the host-fs in
a Ready state, thus enabling it to be deleted.

Partial-Bug: 2122075

Test Plan:
  PASS: Run `host-host-fs-delete <host> ceph --force` and observe the
	components of the host-fs being assigned for force deletion
  PASS: Run `host-stor-delete <host> ceph` and observe no alterations
        from previous behavior
  PASS: Request DELETE /v1/host_fs/%s and ensure it behaves as
        force=False for retrocompatibility

Change-Id: Ia96b1261b75f12fd2fa21dfda019f6a8192c6a85
Signed-off-by: Caio Correa <caio.correa@windriver.com>