Improved agent config update logging
Formerly, it was not possible to see if the agent was receiving config
update requests without debug logging enabled.  This made certain issues
difficult to trace.

This review updates the logging so that it is emitted to the logs with default configuration.

TEST PLAN
PASS: AIO-DX deployed
* No regressions
PASS: Standard+Storage deployed
* No regressions
PASS: AIO-DX, license install
* No regressions
* New logs visible
PASS: Standard+Storage, license install
* No regressions
* New logs visible

Closes-Bug: 2122522
Change-Id: I0f86afbeec5c10a90cf9e9662a2ba2d37afeed9a
Signed-off-by: Joshua Kraitberg <joshua.kraitberg@windriver.com>