Remove dnsmasq reload on system controller OAM update
Since registry.central resolution was removed from the DNS puppet
template, it is no longer necessary to trigger dnsmasq updates (and
restarts) when the system controller OAM addrpool is updated or its
network is created. The related references were removed.

The registry.central dns name is managed through service parameters,
preventing dnsmasq restarts via puppet that could affect subsequent
Ansible tasks while the service was not yet fully ready for DNS
resolution.

Test Plan
=========

[PASS] Perform bootstrap on a DC system and verify correct resolution
of registry.central in the subcloud. Check both the name and address
of registry.central in the dnsmasq configuration files and in the
service-parameter-list table.

[PASS] Perform enrollment on an SX subcloud using the same factory
config parameters for enrollment and verify the correct
registry.central configuration after the process. Also check if DNS
resolution is correct before the unlock.

[PASS] Perform enrollment on an SX subcloud with mgmt reconfiguration
and verify the correct registry.central configuration after the
process. Also check if DNS resolution is correct before the unlock.

[PASS] Perform subcloud rehoming and verify that registry.central is
updated with the new system controller OAM IP. Also check if DNS
resolution is correct.

[PASS] Reconfigure the system controller OAM network and verify that
dnsmasq is not restarted. Ensure registry.central still present in
dnsmasq configuration and listed in the service-parameter-list table.

[PASS] Reconfigure the system controller OAM address pool and verify
that dnsmasq is not restarted. Ensure registry.central still present in
dnsmasq configuration and listed in the service-parameter-list table.

Closes-Bug: 2122452

Change-Id: Ieae31568c6aa7c2ada33a5f47e999d247cb903d8
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>