Improve cert-validation operation for IPsec
The IPsec operation 3 for cert-validation is executed during system
boot and should prevent cases where IPsec certificates are invalid
in the system.

This commit improves cert-validation operation implemented for IPsec
on management network. The IPsec certificate should have its
signature and expiry dates validated during this operation in order
to signalize any need to renew or update this certificate.

Test Plan:
PASS: In a DX system, execute ipsec-client for cert-validation manually
      and observe that the operation completed successfully.
PASS: In a DX system, install and deploy a worker node and observe that
      cert-validation was executed after node unlock.
PASS: Lock and unlock the worker node and observe that the operation 3
      is completed successfully.
PASS: With IPsec certificate expired on worker node, lock and unlock
      this node and observe that IPsec operation 3 was executed and has
      signalized that IPsec certificate is expired. IPsec operation 2
      executed successfully and renewed the IPsec certificate for
      worker node.

Closes-Bug: 2124275

Change-Id: I3337a7958d4c5dcf3e95340c004b619d08bdc650
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>