Set mgmt reconfig flag before database update
During enrollment, a periodic dnsmasq restart may occur between the
database update and the set of the MGMT_NETWORK_RECONFIGURATION_ONGOING
flag, as it is triggered by the sysinv-conductor service restart.

The "update system configurations" task updates the database with the
new management IP and only then sets the flag
MGMT_NETWORK_RECONFIGURATION_ONGOING. If the dnsmasq restart happens
in between, dnsmasq reloads and applies the new mgmt IP too early,
causing controller.internal to point to the new IP before unlock.

This change sets the flag before the database updates to prevent
premature dnsmasq restarts and consequently the incorrect update of
the controller.internal DNS entry to the new mgmt IP.

Test Plan
=========

[PASS] In a system SX, run addrpool modify on mgmt changing the
floating address. Check that
/etc/platform/.mgmt_network_reconfiguration_ongoing is created and
that the flag is set before the database update. Verify this using
sysinv.log and the database entry update timestamp.

[PASS] In a system SX, delete the secondary mgmt addrpool. Verify
it is removed correctly and that
/etc/platform/.mgmt_network_reconfiguration_ongoing is set.

[PASS] In a system SX, create a secondary mgmt network_addrpool.
Check that /etc/platform/.mgmt_network_reconfiguration_ongoing is
set before the database entry is created. Verify this using
sysinv.log and the database operation timestamp.

[PASS] In a system SX, delete the secondary mgmt network_addrpool.
Verify it is removed correctly and that
/etc/platform/.mgmt_network_reconfiguration_ongoing is set.

[PASS] In a system SX, create a mgmt interface-network. Ensure
that /etc/platform/.mgmt_network_reconfiguration_ongoing is set
before the database entry is created. Verify this using sysinv.log
and the database operation timestamp.

[PASS] In a system SX, delete the mgmt interface-network. Ensure
that /etc/platform/.mgmt_network_reconfiguration_ongoing is set
before the database entry is deleted. Verify this using sysinv.log
and the database operation timestamp.

[PASS] Perform enrollment on an SX system with management reconfig. In
the "Update system configurations" task, ensure
/etc/platform/.mgmt_network_reconfiguration_ongoing is set before the
database entry is created. Verify this using sysinv.log and the database
operation timestamp. Check that dnsmasq does not restart after the flag
is set.

[PASS] Perform enrollment on a DX system with management reconfig and
verify that in the "update system configurations" task the reconfig is
skipped with the log "Ignore management network update in duplex".

Closes-bug: 2124204

Change-Id: Ia3091f8907fb465103befe2c2e40cb26767dc17c
Signed-off-by: Ferdinando Terada <Ferdinando.GodoyTerada@windriver.com>