Remove management IP from registry certificate
The system-registry-local-certificate currently contains the management
IP in its Subject Alternative Names, even though this is not necessary
anymore, since there is already the `registry.local` name for this
purpose. This commit removes the management IP from this certificate.

Test Plan:
PASS: build-pkgs & build-image
PASS: Bootstrap an AIO-SX and check that the certificate does not
      contain the management IP.
PASS: Change OAM IP in an AIO-SX, unlock controller and check the
      certificate.
PASS: Change management IP in an AIO-SX, unlock controller and check
      the certificate.
PASS: Perform a successful rehoming after changing the IP and updating
      the subcloud.

Closes-Bug: 2125698
Change-Id: Idc5e925104549c3371a58cd73bf3c6c706422055
Signed-off-by: Rodrigo Tavares <Rodrigo.dosSantosTavares@windriver.com>