Trigger lifecycle action after application recovery
When a rook-ceph application upgrade/downgrade was aborted, the
storage-backend remained in 'updating' state. This occurred because
kube_app was not triggering a lifecycle hook after the application
recovery process completed.

This change adds a call to trigger the app lifecycle after a recovery is
finished, allowing the apps to track this and the rook-ceph app to keep
the storage-backend corretly updated.

TEST PLAN:
- PASS: Build an image and perform a fresh install.
- PASS: Apply the rook-ceph application.
- Downgrade the rook-ceph application and abort the action, and check:
    - PASS: If the storage-backend is updated to 'updating' when the
            abort command is performed.
    - PASS: If the storage-backend is updated to 'applied' when the app
            is recovered.
- PASS: Check if the new lifecycle hook is triggered after an app
        recovery
- PASS: Upload/Apply/Remove metrics-server app

Partial-bug: 2125607

Change-Id: Ib63fbd72779de123aa24654974fb6c94d27c4bc4
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>