Uploaded apps display incorrect progress
If an app is uploaded and it has dependencies (key dependent_apps in
metadata.yaml) that haven't yet been installed on the platform, both
"system application-list" and "system application-show" will show that
the app has missing dependencies.

However, if the user manually installs the dependency for that app, or
even if the dependency is updated to meet the desired version, this
doesn't update the message for apps in the "uploaded" status.

This meant that even if an app's dependency was met, as it happened
after it was uploaded to the platform, the progress message for apps in
the "uploaded" status wouldn't update, still showing that there were
missing dependencies.

This change fixes this by triggering a function at the end of the
entire app apply and update process that checks all "uploaded" apps,
verifies whether the dependency was met, and updates this information
in the database.

Test Plan:
PASS: build-pkgs && build-image
PASS: Fresh install AIO_SX
PASS: Upgrade from STX-10 to master
PASS: upload/apply/delete/remove/update metrics-server app
PASS: If app-1 is successfully applied/updated, but had app-2 that
      depended on app-1 in the uploaded state, the progress column
      message for app-2 is updated to no longer contain the
      dependency on app-1.

Closes-bug: 2126575

Change-Id: I7cbf5ac23788b054b43f8fe5a326a37cb87ddd83
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>