Add flag to skip keystone password update
When a keystone user password is changed, a keystone listener monitor
these changes and may run a custom manifest for certain users.
To better control when the manifest should be executed, this commit
introduces the flag skip_keystone_password_update to selectively
skip the manifest execution. This will be intergrated in the rehoming
and enrollment playbooks to speed up their timings.

Test plan:
  - PASS: Touch the flag and modify a user password. Verify the
          manifest execution was skipped.
  - PASS: Remove the flag and modify a user password. Verify the
          manifest execution was executed successfully.
  - PASS: Rehome a subcloud and verify the behavior is the same as
          before.
  - PASS: Enroll a subcloud and verify the behavior is the same as
          before.

Closes-bug: 2126803

Change-Id: Idd2f97265dea6a8041503f290094bd2f55d984b5
Signed-off-by: Victor Romano <victor.gluzromano@windriver.com>