Set AppArmor profiles statically for IPsec
It is experienced in some environments a behavior where AppArmor
service is enabled before StrongSwan service is started. In the current
implementation, this behavior can lead to a failure as StrongSwan will
be prevented from loading IPsec certificate and storing it in LUKS
filesystem due to AppArmor service being enabled but no AppArmor
profiles for IPsec are configured/applied in the system yet.

Consequently, this behavior results in StrongSwan being prevented to
establish IPsec SAs with other nodes in the environment, leading the
node to a failed state.

This commit generates AppArmor profiles statically for IPsec in system
per the addition of them in the ipsec-auth package files. Whenever
AppArmor service is started, the AppArmor profiles for IPsec will be
applied as well.

Test Plan:
PASS: Build and install ipsec-auth package and observe that AppArmor
      profiles for IPsec (usr.sbin.swanctl and usr.lib.ipsec.charon)
      are stored at /etc/apparmor.d/local directory path.
PASS: Build and deploy an image containing these changes for a DX
      system. Add a worker node with apparmor enabled. Observe that
      AppArmor profiles for IPsec are applied and the node is deployed
      successfully.
PASS: Build and deploy an image containing these changes for a DX
      system. Add a worker node with apparmor disabled. Observe that
      AppArmor profiles for IPsec are configured/stored but not applied
      and node is deployed successfully.
PASS: In the DX+ system, lock worker-0 and disable AppArmor using
      host-update command. Unlock the worker-0 and observe that node
      comes up unlocked available online w/ AppArmor disabled.
PASS: In the DX+ system, lock worker-0 and re-enable AppArmor using
      host-update command. Unlock the worker-0 and observe that node
      comes up unlocked available online w/ AppArmor enabled.

Note: All test plan was executed with the following reverts applied in
      system.
      [1] https://review.opendev.org/c/starlingx/stx-puppet/+/963392
      [2] https://review.opendev.org/c/starlingx/stx-puppet/+/963393

Closes-Bug: 2127083

Change-Id: I8afc5ec5066b6efdecc8f15fc8659de2c4d3ae37
Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>