Add RPC call to audit local registry secrets
As part of rehoming, the keystone password for sysinv is updated for
the new SystemController value. A callback then comes from keystone
to sysinv in which the secrets that contain the password used to
authenticate to the docker registry for each namespace are updated.

However, since the rehoming playbook needs to restart sysinv right
away, a race condition where sysinv restarts before the keystone
callback is completely executed can leave old values in the secrets.

This review adds a RPC method to 'audit_local_registry_secrets',
which updates the credentials in the secrets. This RPC method will
be called from the rehoming ansible playbook, removing the
dependency on the keystone callback.

Test Plan:
PASS: Rehome SX subcloud
      Confirm that the registry secrets were updated

PASS: Enroll SX subcloud

Partial-bug: 2126680

Change-Id: I674111330587c2ecbf966771f4efa2c3b023158f
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>