Robustness improvement for kube-rootca-update
A timing issue may occur when a kube-rootca secret is created but its
crt and key fields are not yet valid when read by subsequent steps.

This commit adds a wait loop after generating or uploading the secret,
until the secret exists and contains valid certificate and key data.
This aims to ensure that subsequent operations can use the secret
without encountering incomplete or missing data.

Additionally, this change replaces the use of the string 'undef' with
`None` when building kube-rootca Puppet parameters. The former is
interpreted by Puppet as a string, instead of as `undef`.

Test Plan:
PASS: build-pkgs & build-image
PASS: Update kube root CA using manual `system` steps in an AIO-SX
      successfully
PASS: Update kube root CA using `sw-manager` strategy in an AIO-SX and
      AIO-DX+Worker successfully
PASS: Run the trust-both-cas phase with an incomplete secret, and make
      sure it fails after 40 seconds
PASS: Run the trust-both-cas phase with an incomplete secret, and make
      sure it completes successfully as soon as the secret is updated
      with the complete fields
PASS: Apply an sw-manager kube-rootca-update-strategy but replace the
      uploaded cert/key in the secret with empty data to simulate the
      timeout issue. Make sure the strategy safely aborts with the
      correct error.

Closes-Bug: 2129759

Change-Id: Ifc66c26f44adcfe1eb195230fa99c9a5c40e2864
Signed-off-by: Rodrigo Tavares <Rodrigo.dosSantosTavares@windriver.com>