Avoid exception when adminep secret is not present
When updating the hieradata, sysinv consults the secret for the
DC adminep ICA (sc-cert/sc-adminep-certificate). When this secret
can't be read, an exception is generated, which in turn prevents
the hieradata to be updated.

In a rehoming race condition, this was causing some files to be
overwritten with outdated data. The next puppet manifest was using
a non updated hieradata, since the exception prevented the update.

This change updates the hieradata with an empty string ('') if
this exception happens. In puppet, there is a verification that
doesn't update the target files when the values are empty, which
prevents the files to be updated with an invalid value.

Test Plan:
PASS: Deployed DC SystemController + SX subcloud.
PASS: Rehomed subcloud.
PASS: Rehomed the dc adminep secret in the subcloud.
      Updated a config to make sysinv run a puppet manifest.
      Checked that the hieradata has empty values for
      platform::config::dccert::params::dc_root_ca_crt and
      platform::config::dccert::params::dc_adminep_crt.
      Checked there is no exception generated.
      Checked that the certificate files were not overwritten.

Partial-bug: 2127875

Change-Id: I1640f5b3803b6f566d77c249553f913c1f27ce86
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>