Limit required admin interface to controller node unlock
It was observed that for standard installations that contain the admin
network, non-controller nodes cannot be unlocked because the
validation is requiring the compute node to have an admin interface
associated with this network. The problem is that admin networks are
only required in controller nodes.

This changes adds this condition to the validation.

Test Plan
=========
[PASS] install a standard setup with the admin network present and
       unlock the compute nodes

Closes-Bug: 2129207

Change-Id: I6a756cf563dbc81e549ad9918913f43ff36e23ea
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>