Generate routes file content in final string format
Tests show that when the number of routes reaches into the thousands,
the application of the Puppet class platform::network::routes can take
several minutes to complete. The primary cause is the use of Puppet's
network_route resource, which is part of the puppet-network plugin
and is responsible for writing the updated
/var/run/network-scripts.puppet/routes file.

This change introduces the generation of route data in the hieradata
format, aligning it with the format used by the networking service's
route files. This approach eliminates the need for the plugin,
improving performance and reducing execution time.


Test Plan
=========
[PASS] install AIO-SX
[PASS] Install 8000 routes and observe the execution time for the
       class platform::network::routes (average of 3s in vbox)
[PASS] Forcibly remove the /etc/network/routes file and create an
       extra route and observe that no errors occur with the existing
       8000 routes and the 8001th is also created
[PASS] Forcibly remove the /etc/network/routes file and lock/unlock
       the node to see that all routes are recreated
[PASS] install an AIO-DX with both nodes unlocked
[PASS] install a Standard with all nodes unlocked


Story: 2011311
Task: 53005


Change-Id: I9e36d5eaeb6feb937d99e5debc571a52341403cc
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>