Force pxeboot interface label generation to ":2"
If the pxeboot network database object for some reason is erased
and recreated it will no longer contain the original ID of 2, obtained
during bootstrap. This creates problems for mtcClient to identify the
correct pxeboot configuration file in /etc/network/interface.d/.

This change fixes the label to ":2", that will create a fixed label
filename.

Test Plan
=========
[PASS] bootstrap an AIO-DX and delete and recreate the pxeboot network
       then procced with the installation and observe no communication
       alarms

Closes-Bug: 2130581

Change-Id: I869164611c9b54701b61b3f751508a5bbd98e248
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>