Consider existing routes when generating for each intf
In [1] the tests added 8000 routes in the mgmt interface.
Since the mgmt interface was the last one created in the database
(done via CLI) its content got processed and configured in the OS.
That is why it wasn't noticed that the code wasn't considering the
routes for the previous processed interfaces.

[1] https://review.opendev.org/c/starlingx/config/+/964795

This change adds that condition.

Test Plan
=========
Done in AIO-SX and AIO-DX
[PASS] adds routes on all interfaces and check their installation at:
       - runtime
       - lock/unlock
[PASS] add routes on all interfaces before the first unlock, after
       bootstrap

Closes-Bug: 2130748

Change-Id: I3d1ce27225258f54f51eb9012f40194eaec69283
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>