Cleanup obsolete out-of-tree parameters
Existing behaviour - by default out-of-tree-drivers
drivers(ice/i40e/iavf) are loaded and switching to in-tree-drivers
is not supported. So, out_of_tree_drivers service parameter and
multi-drivers-switch/out-of-tree-drivers boot parameters which are
used for switching the drivers are now obsolete.

As part of this commit - ocurences of the above mentioned parameters
are being removed

TestPlan:
PASS: build-pkgs
PASS: build-image
PASS: bootstrap
PASS: AIO-SX
      - validated no out_of_tree_drivers service parameter exists
      - validated no out-of-tree-drivers kernel parameter exists
      - validated that out_of_tree_drivers service parameter
        cannot be added.

Story: 2011606
Task: 53160

Change-Id: Ibd5eadcd9179eb29c974cadfa7c5573cfefd10ce
Signed-off-by: Rahul Roshan Kachchap <rahulroshan.kachchap@windriver.com>