Generate GlobalNetworkSet Configuration
This change introduces the use of GlobalNetworkSets to store DC routes
added to the firewall. This approach prevents the creation of large
GlobalNetworkPolicies in the system controller, which previously
occurred due to the high number of managed subclouds. The oversized
policies were causing the kube-API to reject the configuration
because of its size in bytes.

A GlobalNetworkSet is a Calico resource that can hold all subnets to
be used in the source or destination fields of a policy rule. These
sets are linked using a selector, which matches its value against
the labels defined in the GlobalNetworkSet.

Test Plan
=========
[PASS] Install a non-DC AIO-DX dual-stack system, add routes to the
        management interface, and verify that no GlobalNetworkSet
        is created.
[PASS] Install a system controller AIO-DX dual-stack system, add
        routes to the management interface, and verify that a
        GlobalNetworkSet is created.
[PASS] Install a subcloud AIO-SX dual-stack system, add routes to the
        management interface, and verify that a GlobalNetworkSet is
        created.
[PASS] Install a subcloud AIO-SX dual-stack system, add routes to the
        admin interface, and verify that a GlobalNetworkSet is
        created.
[PASS] Validate firewall functionality between the subcloud and system
        controller.

Story: 2011311
Task: 53249

Change-Id: I3056847a01b7c0d6bb098b8e848735d13afbe9d7
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>