Trigger rook migration via sysinv API
This allows the migration from Ceph bare metal to Rook Ceph via sysinv
API.

The system command to trigger is:

  system storage-backend-modify <uuid> migration_type=<migration-type>

The API request is:

  PATCH /v1/storage_backend/<uuid>
  [{
    "path": "/capabilities",
    "value": "{\"migration_type\": \"<migration-type>\"}",
    "op": "replace"
  }]

The possible "migration_type" is only redeploy for now,
but in-service and export-import can be supported in the future.

Additionally, a fix was implemented to storage_backend controller
patch() when redirecting to storage_ceph_rook. It was trying to redirect
as storage_ceph-rook() and returning 500 error code.

Test Plan:
  PASS: Test triggering migration via storage-backend-modify
  PASS: Ensure migration process running is non-blocking
  PASS: Test passing invalid migration_type gives proper error
  PASS: Ensure storage_backend patch() redirects to storage_ceph_rook

Story: 2011640
Task: 53537

Depends-on: https://review.opendev.org/c/starlingx/ansible-playbooks/+/971359

Change-Id: I1187a2e0d633bfd35c3fef2f74b892274bb2b1d5
Signed-off-by: Ítalo Vieira <italo.gomesvieira@windriver.com>