Skip application auditing via service parameter
With this change, the manual operation of disabling or enabling
auditing will now be done via a service parameter and no longer via
sysinv.conf.

This new approach was adopted due to the lack of persistence via
sysinv.conf. If the user changed sysinv.conf with the new value, but
then performed a swact or even a lock and unlock, the value would be
reset to the default.

With the service parameter, the value is persisted in the database,
eliminating the possibility of a reset after a swact or unlock.

Test Plan:
PASS: build-pkgs && build-image
PASS: deploy AIO-SX and AIO-DX

Simplex:
PASS: deploying a simplex, the value of the k8s_application_audit
      parameter is set to enable by default.
PASS: After running the command "system service-parameter-modify
      kubernetes config k8s_application_audit=disabled" the next audit
      interaction, which occurs every 1 minute, is skipped..
PASS: After running the command "system service-parameter-modify
      kubernetes config k8s_application_audit=enabled" the next audit
      interaction has already been re-established.
PASS: After lock and unlock of host, the parameter value remains the
      same, without any changes.

Duplex:
PASS: After running the command "system service-parameter-modify
      kubernetes config k8s_application_audit=disabled" and performing
      a swact of the host, the value of k8s_application_audit remains
      disabled.

Story: 2011242
Task: 53299

Change-Id: Ifc13048d23b26c0f965ad6a18f70208395baa9c4
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>